feat(web): support x402 multi-accepts payment via A2A_X402_ACCEPTS
