+++ apps/web/.env.example
@@ -19,13 +19,15 @@ FAUCET_ADMIN_ADDRESS=
 
 DATABASE_URL=
 
+# x402 payment facilitator endpoint
+X402_FACILITATOR_URL=https://x402.org/facilitator # public testnet facilitator endpoint
+
 # x402 payment extension (Merchant Agent config)
 # A2A_X402_PAY_TO — wallet address that receives payments (required to enable x402 flow)
-# A2A_X402_NETWORK — blockchain network (default: base-sepolia)
-# A2A_X402_ASSET   — ERC-20 token contract address (default: USDC on the selected network)
-# A2A_X402_MAX_AMOUNT — required payment amount in token base units (default: 1000000 = 1 USDC)
+# A2A_X402_ACCEPTS — JSON array of accepted payment options (network/asset/maxAmount per entry).
+#   Each entry: { "network": "<network>", "asset": "<token-address>", "maxAmount": "<amount-in-base-units>" }
+#   Example (two networks):
+#     A2A_X402_ACCEPTS=[{"network":"base-sepolia","asset":"0x036CbD53842c5426634e7929541eC2318f3dCF7e","maxAmount":"1000"},{"network":"base","asset":"0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913","maxAmount":"1000"}]
+
 A2A_X402_PAY_TO=
-A2A_X402_NETWORK=base-sepolia
-A2A_X402_ASSET=0x036CbD53842c5426634e7929541eC2318f3dCF7e # USDC on base-sepolia
-A2A_X402_MAX_AMOUNT=1000 # 0.001 USDC
-X402_FACILITATOR_URL=https://x402.org/facilitator # public testnet facilitator endpoint
+A2A_X402_ACCEPTS=
