+++ packages/x402/tsconfig.json
@@ -8,7 +8,8 @@
     "outDir": "./dist",
     "declaration": true,
     "isolatedModules": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "types": ["node"]
   },
   "include": ["src"],
   "exclude": ["node_modules"]
